src/requirements.txt
flask
flask_cors
flask-oauth
flask_compress
flask_login
pymongo
requests
python-dateutil
schedule
python-dotenv
pytz
tqdm
nltk==3.2.5
shapely
haversine
fuzzywuzzy
unidecode
psycopg2
Flask-SQLAlchemy

# Machine Learning
scikit-learn
numpy
scipy
pandas
